requirements.txt
# Core Dependencies
yfinance>=0.2.66
pandas>=2.3.3
requests>=2.32.5
beautifulsoup4>=4.14.2
numpy>=1.24.0

# Web Framework
fastapi>=0.120.1
uvicorn>=0.38.0
gunicorn>=23.0.0
flask>=3.1.2

# System & Utilities
psutil>=7.1.2
schedule>=1.2.2
pydantic>=2.12.3
websockets>=15.0.1
curl-cffi>=0.13.0
holidays>=0.83
python-dotenv>=1.2.1
lxml>=4.9.0

# Sentiment Analysis
textblob>=0.17.1
